+++ src/components/App.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import logo from "./logo.svg";
-import "./App.css";
+import logo from "../logo.svg";
+import "../static/css/App.css";
 import Navigation from './Navigation';
 
 const App = () => {
